requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
bitcoin==1.1.42
celery==3.1.25
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
click==6.7
cytoolz==0.8.2
Django==1.11.5
django-solo==1.1.2
eth-testrpc==1.3.0
ethereum==1.6.1
ethereum-abi-utils==0.4.1
ethereum-tester==0.1.0a6
ethereum-utils==0.4.1
idna==2.6
json-rpc==1.10.3
kombu==3.0.37
pbkdf2==1.3
pluggy==0.5.2
py==1.4.34
pycparser==2.18
pycryptodome==3.4.7
pyethash==0.1.27
pylru==1.0.9
pysha3==1.0.2
pytz==2017.2
PyYAML==3.12
repoze.lru==0.7
requests==2.18.4
rlp==0.6.0
scrypt==0.8.0
secp256k1==0.13.2
semantic-version==2.6.0
toolz==0.8.2
tox==2.8.2
urllib3==1.22
virtualenv==15.1.0
web3==3.15.0
Werkzeug==0.12.2

